fix warning message : value is trying to be set on a copy of a slice from a DataFrame and an error where f.seek posNew expect integer but got float

+++ taxonomy.py
@@ -83,7 +83,7 @@ def acc2taxid( acc ):
 			
 			while( acc != accCur and start < end ):
 				
-				posNew = (end+start)/2
+				posNew = int((end+start)/2)
 				
 				f.seek( posNew )
